// src/app/models/response.model.ts
export class Response {
  Segment?: string;
  Standing?: string;
  Espace_a_traiter?: string;
  Style?: string;
  Forme?: string;
  Surface_de_pose?: string;
  Mode_de_pose?: string;
  Combinaisons?: string;
  Choix1?: string;
  Choix2?: string;
  Choix3?: string;
  Type_de_projet?: string;
  Materiaux_de_surface_de_pose?: string;
  Finition?: string;
};